Handle FD62NP(N) by it own EEP
Background: used to be handled by generic dim action, but the FD62 has features with I am using often and which in turn are not being covered by TF-13-07 (LM).

So I made a EEP for the FD62 only and gave this EEP the number TF-13-99. 99 to prevent merge clashes.