Provide manual device creation entries for FD62 and FL62
I use this a lot. Every time I remove an FD62/FL62 in ioBroker, or make a new ioBroker setup, I do not have do (re)learn the USB300 into the Eltako actors, they already know the USB300. So I just create the device in ioBroker with the known device-id and immediately it is working.